chore: move fetching realtime schedule to crud file
