+++ src/translink/crud.py
@@ -1,12 +1,22 @@
+import asyncio
 import io
 import zipfile
 from datetime import datetime
-from typing import cast
+from typing import Any, cast
 
+import httpx
 import pandas as pd
+from google.transit import gtfs_realtime_pb2
 from httpx import AsyncClient
 
+from config import settings
 from constants import TZ_INFO
+from translink.models import BusScheduleEntry, BusStatus
+from translink.types import FeedMessage
+
+REALTIME_URL = "https://gtfsapi.translink.ca/v3/gtfsrealtime"
+POSITION_URL = "https://gtfsapi.translink.ca/v3/gtfsposition"
+
 
 # Taken from the static data.
 # Key: Route ID
@@ -19,6 +29,18 @@ BUS_DATA = {
     "6658": (1, "1875", "145"),  # Production
     "37807": (1, "3129", "R5"),  # Hastings
 }
+
+
+def get_bus_status_string(status: int) -> BusStatus:
+    match status:
+        case 0:
+            return BusStatus.INCOMING_AT
+        case 1:
+            return BusStatus.STOPPED_AT
+        case 2:
+            return BusStatus.IN_TRANSIT_TO
+        case _:
+            raise ValueError(f"Unknown bus status: {status}")
 
 
 def _gtfs_time_to_seconds(time_str: str) -> int:
@@ -118,3 +140,73 @@ def get_next_departures(schedule: pd.DataFrame, n: int = 3) -> pd.DataFrame:
         .head(n)
         .sort_values(["route_id", "departure_seconds"])
     )
+
+
+async def fetch_feed(client: httpx.AsyncClient, url: str, params: dict[str, Any]):
+    response = await client.get(url, params=params)
+    feed = cast(FeedMessage, gtfs_realtime_pb2.FeedMessage())  # pyright: ignore[reportAttributeAccessIssue]
+    feed.ParseFromString(response.content)
+    return feed
+
+
+async def fetch_realtime_schedule(client: AsyncClient) -> list[BusScheduleEntry]:
+    """
+    Gets the real-time bus schedule from the TransLink GTFS Realtime API.
+    """
+    # FeedMessage is generated at runtime, so the type checker can't find this function
+    params = {"apikey": settings.translink_api_key}
+
+    trip_feed, position_feed = await asyncio.gather(
+        fetch_feed(client, REALTIME_URL, params=params),
+        fetch_feed(client, POSITION_URL, params=params),
+    )
+
+    # Filter for the relevant vehicles
+    vehicle_positions = {}
+    for entity in position_feed.entity:
+        if not entity.HasField("vehicle"):
+            continue
+
+        trip = entity.vehicle.trip
+        if trip.route_id not in BUS_DATA or trip.direction_id != BUS_DATA[trip.route_id][0]:
+            continue
+        vehicle_positions[trip.trip_id] = entity.vehicle
+
+    # Go through all of the trips occurring and filter for the relevant ones,
+    # then determine which ones are delayed.
+    result: list[BusScheduleEntry] = []
+    for entity in trip_feed.entity:
+        if not entity.HasField("trip_update"):
+            continue
+
+        trip_update = entity.trip_update
+        trip = trip_update.trip
+        bus_data = BUS_DATA.get(trip.route_id)
+
+        if bus_data is None or trip.direction_id != bus_data[0]:
+            continue
+
+        _, stop_id, route_number = bus_data
+        stop = next((s for s in trip_update.stop_time_update if s.stop_id == stop_id), None)
+        if stop is None:
+            continue
+
+        scheduled_time = stop.departure.time - stop.departure.delay
+
+        trip_id = trip.trip_id
+        status = BusStatus.INCOMING_AT
+        vehicle = vehicle_positions.get(trip_id)
+        if vehicle:
+            status = get_bus_status_string(vehicle.current_status)
+
+        result.append(
+            BusScheduleEntry(
+                route_number=route_number,
+                scheduled_departure_time=scheduled_time,
+                realtime_time=stop.departure.time,
+                delay_seconds=stop.departure.delay,
+                status=status,
+            )
+        )
+
+    return result
